+++ crates/cosmos/cosmos-chain-components/src/components/client.rs
@@ -137,8 +137,6 @@ pub use hermes_relayer_components::chain::traits::types::update_client::UpdateCl
 
 use crate::components::delegate::DelegateCosmosChainComponents;
 use crate::impls::channel::init_channel_options::ProvideCosmosInitChannelOptionsType;
-use crate::impls::client::create_client_payload::BuildCreateClientPayloadWithChainHandle;
-use crate::impls::client::update_client_payload::BuildUpdateClientPayloadWithChainHandle;
 use crate::impls::connection::init_connection_options::ProvideCosmosInitConnectionOptionsType;
 use crate::impls::events::ProvideCosmosEvents;
 use crate::impls::packet::packet_fields::CosmosPacketFieldReader;
@@ -168,7 +166,6 @@ use crate::impls::queries::write_ack_event::QueryWriteAckEventFromChainHandle;
 use crate::impls::types::chain::ProvideCosmosChainTypes;
 use crate::impls::types::client_state::ProvideAnyRawClientState;
 use crate::impls::types::consensus_state::ProvideAnyRawConsensusState;
-use crate::impls::types::create_client_options::ProvideCosmosCreateClientSettings;
 use crate::impls::types::payload::ProvideCosmosPayloadTypes;
 pub use crate::traits::abci_query::AbciQuerierComponent;
 
@@ -219,8 +216,6 @@ define_components! {
         ]:
             ProvideCosmosEvents,
         [
-            CreateClientPayloadTypeComponent,
-            UpdateClientPayloadTypeComponent,
             ConnectionOpenInitPayloadTypeComponent,
             ConnectionOpenTryPayloadTypeComponent,
             ConnectionOpenAckPayloadTypeComponent,
@@ -254,12 +249,6 @@ define_components! {
             RawConsensusStateWithProofsQuerierComponent,
         ]:
             QueryCosmosConsensusStateFromAbci,
-        CreateClientPayloadOptionsTypeComponent:
-            ProvideCosmosCreateClientSettings,
-        CreateClientPayloadBuilderComponent:
-            BuildCreateClientPayloadWithChainHandle,
-        UpdateClientPayloadBuilderComponent:
-            BuildUpdateClientPayloadWithChainHandle,
         CounterpartyChainIdQuerierComponent:
             QueryChainIdWithChainHandle,
 
@@ -347,6 +336,10 @@ define_components! {
             ConsensusStateTypeComponent,
             ConsensusStateFieldComponent,
 
+            CreateClientPayloadTypeComponent,
+            UpdateClientPayloadTypeComponent,
+            CreateClientPayloadOptionsTypeComponent,
+
             ConsensusStateHeightsQuerierComponent,
             CounterpartyMessageHeightGetterComponent,
 
@@ -354,6 +347,9 @@ define_components! {
 
             CreateClientMessageBuilderComponent,
             CreateClientMessageOptionsTypeComponent,
+
+            CreateClientPayloadBuilderComponent,
+            UpdateClientPayloadBuilderComponent,
 
             ClientStateQuerierComponent,
             ClientStateWithProofsQuerierComponent,